Methods

Index

AstroIC.Rd_from_RHI

AstroIC.addpos

AstroIC.addvel

AstroIC.exponential_cdf

AstroIC.exponential_cdf_inv

AstroIC.exponential_pdf

AstroIC.exponential_pdf_hole

AstroIC.freefall_velocity_acc

AstroIC.generate

AstroIC.generate

AstroIC.generate

AstroIC.generate

AstroIC.generate

AstroIC.generate

AstroIC.generate_milkyway_baryon_particles

AstroIC.gridpoints

AstroIC.helio2xyz

AstroIC.load_MW_RC_DS_W21

AstroIC.load_MW_RC_Eilers2019

AstroIC.load_MW_RC_Mroz2019

AstroIC.load_MW_RC_stddev_W21

AstroIC.load_SPARC_LTGs_RC

AstroIC.load_SPARC_LTGs_data

AstroIC.load_SPARC_Xray_ETGs_data

AstroIC.load_SPARC_rotating_ETGs_RC

AstroIC.load_SPARC_rotating_ETGs_data

AstroIC.load_SPARC_rotating_ETGs_rotmod

AstroIC.load_data_MW_satellites

AstroIC.load_data_UFDs

AstroIC.load_li2018_SPARC

AstroIC.load_massive_dwarf_Baryon_RC

AstroIC.load_massive_dwarf_CO_RC

AstroIC.load_massive_dwarf_DM_RC

AstroIC.minimum_func

AstroIC.rand_pos_2d

AstroIC.rand_pos_3d

AstroIC.rejection_sampling

AstroIC.rejection_sampling

AstroIC.rotational_velocity

AstroIC.rotational_velocity_acc

AstroIC.sech2_cdf

AstroIC.sech2_cdf_inv

AstroIC.sech2_pdf

AstroIC.setpos

AstroIC.setvel

AstroIC.solarsystem

AstroIC.vmean

AstroIC.vmean2

Initial-condition generation

AstroIC.generate — Function

function generate(::InitialConditionConfig, units; kw...)

Generate initial conditions in units

Usable IC configs

PlummerStarCluster

GasCloud

ExponentialDisc

Common keywords

constants

AstroIC.generate — Method

generate(config::PlummerStarCluster; ...) -> StructArray
generate(
    config::PlummerStarCluster,
    units;
    constants,
    MaxRadius
) -> StructArray

Generate initial conditions of Plummer model

Automatically promote to Measurement if the parameters in config has Measurement type

Keywords

MaxRadius: resample particles outside the interested radius. Default is 5 * VirialRadius. Set to zero to avoid cutting off.

Common keywords

constants

AstroIC.generate — Method

generate(config::ExponentialDisc; ...) -> StructArray
generate(
    config::ExponentialDisc,
    units;
    RotationCurve,
    MaxRadius,
    MaxHeight,
    k,
    bc,
    rotational_ratio
) -> StructArray

AstroIC.generate — Method

generate(config::Bulge; ...) -> StructArray
generate(
    config::Bulge,
    units;
    RotationCurve,
    MaxRadius,
    MaxHeight
) -> StructArray

AstroIC.generate — Method

function generate(config::GasCloud, units = uAstro; kw...)

Generate gas cloud sampled from a Cartesian grid

Keywords

Common keywords

constants

AstroIC.generate — Method

generate(config::SphericalSystem; ...) -> StructArray
generate(
    config::SphericalSystem,
    units;
    MaxRadius
) -> StructArray

AstroIC.solarsystem — Function

function solarsystem(date::DateTime)
function solarsystem(date::Real)

Generate initial conditions of Solar System at desinated date

Particle-system manipulation

AstroIC.setpos — Function

function setpos(data::Union{Array, StructArray}, pos::AbstractPoint)

Set system center (middle value) to pos

AstroIC.setvel — Function

function setvel(data::Union{Array, StructArray}, vel::AbstractPoint)

Set system velocity (mass weighted average) to vel

AstroIC.addpos — Function

function addpos(data::Array, pos::AbstractPoint)
function addpos(data::StructArray, pos::AbstractPoint)

Add pos to :Pos of all particles

AstroIC.addvel — Function

function addvel(data::Array, vel::AbstractPoint)
function addvel(data::StructArray, vel::AbstractPoint)

Add vel to :Vel of all particles

Geometry & physics helpers

AstroIC.gridpoints — Function

function gridpoints(R::Number, Nx::Int, Ny::Int, Nz::Int)

Return a Tuple (x, y, z) of Cartesian grid points.

R: half of the sidelength of box.

Nx, Ny, Nz: number of cells in each direction

AstroIC.helio2xyz — Function

function helio2xyz(jd, num)

Convert heliocentric coordinates (in unit AU) to Cartesian coordinates in unit m

Returns a PhysicalParticles::PVector

AstroIC.vmean — Function

vmean(T::Number, mass::Number, units = uAstro)

Thermal (Maxwell–Boltzmann) mean speed $\sqrt{8 k_B T / (\pi m)}$ of a particle of mass mass at temperature T, converted to the velocity unit of units.

Arguments

T: temperature (with units, e.g. 300u"K")

mass: particle mass (with units, e.g. Constant().m_p or 1.0u"Msun")

units: target unit system (default uAstro); must support getuVel

Returns

A velocity in the velocity unit of units (e.g. kpc/Gyr for uAstro, m/s for uSI).

Example

v = vmean(300u"K", Constant().m_p, uSI)   # ≈ 1.15 km/s for a proton

AstroIC.vmean2 — Function

vmean2(T::Number, mass::Number, units = uAstro)

Squared thermal mean speed $8 k_B T / (\pi m)$ of a particle of mass mass at temperature T, converted to the squared-velocity unit of units.

See also vmean.

Data loaders

AstroIC.load_data_MW_satellites — Function

load_data_MW_satellites() -> DataFrames.DataFrame

Load src/data/MW_satellites.csv with units in Galactocentric coordinates.

Reference: Table B.1 and Table B.2 from battaglia2022gaia_Astron. Astrophys. - Gaia early DR3 systemic motions of local group dwarf galaxies and orbital properties with a massive Large Magellanic Cloud

AstroIC.load_data_UFDs — Function

load_data_UFDs() -> DataFrames.DataFrame

Load src/data/UFDs.csv with units

Reference: Table 1 and Table A.1 of hayashi2023dark_Astrophys. J. - Dark matter halo properties of the galactic dwarf satellites implication for chemo-dynamical evolution

AstroIC.load_SPARC_LTGs_RC — Function

load_SPARC_LTGs_RC() -> DataFrames.DataFrame

AstroIC.load_SPARC_LTGs_data — Function

load_SPARC_LTGs_data(; exclude_list, prior, model) -> Any

AstroIC.load_li2018_SPARC — Function

load_li2018_SPARC()

Load the SPARC Radial Acceleration Relation (RAR) table from Li et al. (2018) Li2018_SPARC_RAR.txt.

Returns a DataFrame with the per-galaxy radial-acceleration relation measurements (disk / bulge decompositions, distance, inclination, …). Several columns (LMR_disk, LMR_bulge, D, Inc) are parsed as Measurement{Float64} so that uncertainties propagate.

AstroIC.load_SPARC_Xray_ETGs_data — Function

load_SPARC_Xray_ETGs_data() -> DataFrames.DataFrame

AstroIC.load_SPARC_rotating_ETGs_data — Function

load_SPARC_rotating_ETGs_data() -> DataFrames.DataFrame

AstroIC.load_SPARC_rotating_ETGs_RC — Function

load_SPARC_rotating_ETGs_RC(
    GalaxyName::AbstractString,
    mode::Symbol;
    folder
) -> DataFrames.DataFrame

AstroIC.load_SPARC_rotating_ETGs_rotmod — Function

load_SPARC_rotating_ETGs_rotmod(
    GalaxyName::AbstractString;
    folder
) -> DataFrames.DataFrame

AstroIC.load_MW_RC_Eilers2019 — Function

load_MW_RC_Eilers2019() -> DataFrames.DataFrame
load_MW_RC_Eilers2019(filename) -> DataFrames.DataFrame

AstroIC.load_MW_RC_Mroz2019 — Function

load_MW_RC_Mroz2019() -> DataFrames.DataFrame
load_MW_RC_Mroz2019(filename) -> DataFrames.DataFrame

AstroIC.load_MW_RC_stddev_W21 — Function

load_MW_RC_stddev_W21() -> DataFrames.DataFrame
load_MW_RC_stddev_W21(filename) -> DataFrames.DataFrame

AstroIC.load_MW_RC_DS_W21 — Function

load_MW_RC_DS_W21() -> DataFrames.DataFrame
load_MW_RC_DS_W21(filename) -> DataFrames.DataFrame

AstroIC.load_massive_dwarf_CO_RC — Function

load_massive_dwarf_CO_RC(
    galaxy::String
) -> DataFrames.DataFrame

Load CO RC data for massive dwarfs from Cooke2022

Arguments:

galaxy::String: Name of the dwarf galaxy

Returns:

DataFrame with CO RC data including calculated velocity errors

AstroIC.load_massive_dwarf_DM_RC — Function

load_massive_dwarf_DM_RC(
    galaxy::String
) -> DataFrames.DataFrame

Load DM RC data for massive dwarfs from Cooke2022

Arguments:

galaxy::String: Name of the dwarf galaxy

Returns:

DataFrame with DM RC data including calculated velocity errors

AstroIC.load_massive_dwarf_Baryon_RC — Function

load_massive_dwarf_Baryon_RC(
    galaxy::String
) -> DataFrames.DataFrame

Load Baryon RC data for massive dwarfs from Cooke2022

Arguments:

galaxy::String: Name of the dwarf galaxy

Returns:

DataFrame with Baryon RC data including copied velocity column

AstroIC.generate_milkyway_baryon_particles — Function

generate_milkyway_baryon_particles(Np)

Generate Milky Way baryon particles (bulge, thin/thick discs, HI/HII gas).

Internal helpers

These functions are not part of the public API but are used internally by the generators above. They are listed here so that Documenter does not complain about undocumented docstrings, and so that interested users can find them.

AstroIC.Rd_from_RHI — Method

Return Rd in kpc. RHI in kpc

AstroIC.exponential_cdf — Method

exponential_cdf(beta, x) -> Any

Ref: https://en.wikipedia.org/wiki/Exponential_distribution

AstroIC.exponential_cdf_inv — Method

exponential_cdf_inv(beta, y) -> Any

Ref: https://en.wikipedia.org/wiki/Exponential_distribution

AstroIC.exponential_pdf — Method

exponential_pdf(beta, x) -> Any

Ref: https://en.wikipedia.org/wiki/Exponential_distribution

AstroIC.exponential_pdf_hole — Method

exponential_pdf_hole(beta, Rm, x) -> Any

AstroIC.freefall_velocity_acc — Method

freefall_velocity_acc(x, y, z, a) -> Any

AstroIC.minimum_func — Method

minimum_func(
    target::Function,
    ic
) -> Tuple{Any, AbstractArray}

Use Optim.jl to find the minimum value of target function.

AstroIC.rand_pos_2d — Method

rand_pos_2d(r::Number) -> PVector2D

Generate PVector2D from radius

AstroIC.rand_pos_3d — Method

rand_pos_3d(r::Number) -> Union{PVector, PVector2D}

Generate PVector from radius

AstroIC.rejection_sampling — Method

rejection_sampling(x, pdf_array, rMax, NumSamples) -> Any

Sample from the discrete pdf using rejection method

Attentions

Data must be unitless, and have no NaN

x should start from 0

the maximum of pdf_array would be the width of sampling box

AstroIC.rejection_sampling — Method

rejection_sampling(pdf::Function, rMax, NumSamples) -> Any

Sample from the discrete pdf using rejection method

Attentions

Data must be unitless, and have no NaN

x should start from 0

the maximum of pdf would be the width of sampling box

AstroIC.rotational_velocity — Function

rotational_velocity(x, y, v) -> Any
rotational_velocity(x, y, v, rot_ratio) -> Any

Rotate along the positive z-axis.

rot_ratio: rotational motion v.s. random motion. If equals 1, only rotational component

AstroIC.rotational_velocity_acc — Function

rotational_velocity_acc(x, y, z, a) -> Any
rotational_velocity_acc(x, y, z, a, rot_ratio) -> Any

Rotate along the positive z-axis.

ratio: rotational motion v.s. random motion. If equals 1, only rotational component

AstroIC.sech2_cdf — Method

sech2_cdf(x) -> Any

AstroIC.sech2_cdf_inv — Method

sech2_cdf_inv(u) -> Any

AstroIC.sech2_pdf — Method

sech2_pdf(x) -> Any